requirements.txt
Flask==0.10.1
Pillow==2.9.0
lxml==3.4.2